capitalize first char on name & code order MM-442
