+++ services/UserService.js
@@ -1,5 +1,4 @@
-const fs = require('fs');
-const sanitize = require('mongo-sanitize');
+/* eslint-disable no-underscore-dangle */
 const User = require('../models/User');
 const Messages = require('../constants/Messages');
 const PhoneNormalization = require('./utils/PhoneNormalization');
@@ -19,6 +18,7 @@ const {
   missingPath,
   missingId,
 } = require('../constants/serviceErrors');
+const { createImage } = require('./utils/creatreImate');
 
 const { DID_NOT_FOUND } = Messages.VALIDATION;
 
@@ -373,12 +373,7 @@ module.exports.saveImage = async function saveImage(did, contentType, path) {
     if (!user) return Promise.reject(Messages.USER.ERR.GET);
 
     // Crear imagen
-    const cleanedPath = sanitize(path);
-    const image = fs.readFileSync(cleanedPath);
-    const encodedImage = image.toString('base64');
-    const buffer = Buffer.from(encodedImage, 'base64');
-
-    const { _id } = await Image.generate(buffer, contentType);
+    const _id = await createImage(path, contentType);
 
     // Actualizar imagen del usuario
     await user.updateImage(_id);
